Moved RootStack into its own file inside of navigation folder. Also slight change to style of photo modal, now black background around modal image
